+++ src/main/java/net/geforcemods/securitycraft/datagen/RecipeGenerator.java
@@ -229,8 +229,8 @@ public class RecipeGenerator extends RecipeProvider {
 		.save(consumer);
 		//don't change these to reinforced, because the block reinforcer needs a laser block!!!
 		ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, SCContent.LASER_BLOCK.get())
-		.pattern("SSS")
-		.pattern("SRS")
+		.pattern("SGS")
+		.pattern("GRG")
 		.pattern("SGS")
 		.define('S', ItemTags.STONE_CRAFTING_MATERIALS)
 		.define('R', Tags.Items.STORAGE_BLOCKS_REDSTONE)
@@ -238,6 +238,14 @@ public class RecipeGenerator extends RecipeProvider {
 		.unlockedBy("has_stone", has(ItemTags.STONE_CRAFTING_MATERIALS))
 		.save(consumer);
 		//k you can change again :)
+		ShapedRecipeBuilder.shaped(RecipeCategory.MISC, SCContent.LENS.get(), 6)
+		.group("securitycraft:lens")
+		.pattern(" P")
+		.pattern("P ")
+		.pattern(" P")
+		.define('P', SCContent.REINFORCED_GLASS_PANE.get())
+		.unlockedBy("has_glass_pane", has(SCTags.Items.REINFORCED_GLASS_PANES))
+		.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.MISC, SCContent.MINE.get(), 3)
 		.pattern(" I ")
 		.pattern("IGI")